+++ be/src/storage/CMakeLists.txt
@@ -213,9 +213,10 @@ set(STORAGE_FILES
     lake/cloud_native_index_compaction_task.cpp
     lake/metacache.cpp
     lake/lake_primary_key_recover.cpp
-    lake/load_spill_block_manager.cpp
     lake/spill_mem_table_sink.cpp
     lake/tablet_retain_info.cpp
+    load_spill_block_manager.cpp
+    load_chunk_spiller.cpp
     sstable/block_builder.cpp
     sstable/block.cpp
     sstable/coding.cpp
