+++ be/src/storage/lake/async_delta_writer.cpp
@@ -22,7 +22,7 @@
 
 #include "common/compiler_util.h"
 #include "storage/lake/delta_writer.h"
-#include "storage/lake/load_spill_block_manager.h"
+#include "storage/load_spill_block_manager.h"
 #include "storage/storage_engine.h"
 #include "testutil/sync_point.h"
 #include "util/stack_trace_mutex.h"
